Renamed package names to lowercase.
