--- src/main/java/Entities/User.java
@@ -1,5 +1,0 @@
-package Entities;
-
-public class User
-{
-}
