+++ src/main/java/entities/User.java
@@ -1,0 +1,17 @@
+package entities;
+
+/** Represents a user.
+ * @author Kevin Liu
+ * @version 1.0
+ * @since 1.0
+ */
+public class User
+{
+    int id;
+    UserProfile userprofile;
+    String status;
+    List <Chat> chats
+    FriendsList friendsList
+
++ addChat(Chat chat)
+}
