+++ src/main/java/usecases/UseCases.md
