+++ src/main/java/usecases/profile_changes/ChangeProfileInteractor.java
@@ -1,4 +1,4 @@
-package UseCases.ProfileChanges;
+package usecases.profile_changes;
 
 public class ChangeProfileInteractor {
 }
